+++ DirectShowFilters/StreamingServer/Source/TsMPEG1or2DemuxedServerMediaSubsession.h
@@ -25,7 +25,7 @@ private:
   virtual ~TsMPEG1or2DemuxedServerMediaSubsession();
 
 private: // redefined virtual functions
-  virtual void seekStreamSource(FramedSource* inputSource, float seekNPT);
+  virtual void seekStreamSource(FramedSource* inputSource, double seekNPT);
   virtual FramedSource* createNewStreamSource(unsigned clientSessionId,
 					      unsigned& estBitrate);
   virtual RTPSink* createNewRTPSink(Groupsock* rtpGroupsock,
