+++ DirectShowFilters/StreamingServer/Source/TsMPEG2TransportFileServerMediaSubsession.h
@@ -20,7 +20,7 @@ protected:
   virtual __int64 filelength() const;
 
 private: // redefined virtual functions
-	virtual void seekStreamSource(FramedSource* inputSource, float seekNPT);
+	virtual void seekStreamSource(FramedSource* inputSource, double seekNPT);
   virtual FramedSource* createNewStreamSource(unsigned clientSessionId,
 					      unsigned& estBitrate);
   virtual RTPSink* createNewRTPSink(Groupsock* rtpGroupsock,
